Refactoring Billing DB structures.
