+++ db/python/layers/__init__.py
@@ -1,8 +1,7 @@
 from db.python.layers.analysis import AnalysisLayer
 from db.python.layers.assay import AssayLayer
 from db.python.layers.base import BaseLayer
-from db.python.layers.billing_db import BillingDb
-from db.python.layers.billing_layer import BillingLayer
+from db.python.layers.billing import BillingLayer
 from db.python.layers.family import FamilyLayer
 from db.python.layers.participant import ParticipantLayer
 from db.python.layers.sample import SampleLayer
